kotlin: implement type inference
Fixes https://github.com/adsharma/py2many/issues/16

+++ tests/expected/binit.kt
@@ -7,7 +7,7 @@ fun bisect_right(data_: Array<Int>, item: Int): Int {
         if (item < data_[middle]) {
             high = middle
         } else {
-            low = (middle + 1)
+            low = (middle.toInt() + 1)
         }
     }
     return low
